rpk: move cmd/container to its own package
