+++ src/go/rpk/pkg/cli/cmd/container/purge.go
@@ -19,7 +19,7 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
-func Purge() *cobra.Command {
+func newPurgeCommand() *cobra.Command {
 	command := &cobra.Command{
 		Use:   "purge",
 		Short: "Stop and remove an existing local container cluster's data",
